requirements.txt
Telethon==1.41.2
python-dotenv==1.1.1
